refactored vector store component
